requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
streamlit==1.27.0
langchain==0.0.194
milvus
youtube-transcript-api==0.6.0
openai==0.27.6
python-dotenv==1.0.0
chromadb
tiktoken